Valamis 2.6.1

+++ valamis-certificate/src/main/scala/com/arcusys/valamis/certificate/model/goal/ActivityGoal.scala
@@ -3,7 +3,7 @@ package com.arcusys.valamis.certificate.model.goal
 import com.arcusys.valamis.model.PeriodTypes.PeriodType
 
 case class ActivityGoal(
-  certificateId: Int,
+  certificateId: Long,
   activityName: String,
   count: Int,
   periodValue: Int,
